% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geraGraficosMedias.R
\name{geraGraficosMedias}
\alias{geraGraficosMedias}
\title{gera grafico medias}
\usage{
geraGraficosMedias(media_nt_fg_d1_pt_comp, media_nt_fg_d1_ct_comp,
  media_nt_fg_d2_pt_comp, media_nt_fg_d2_ct_comp, media_nt_ce_d1_comp,
  media_nt_ce_d2_comp, media_nt_ce_d3_comp)
}
\arguments{
\item{media_nt_fg_d1_pt_comp:}{Media de Nota de Lingua Portuguesa da questao 1 da parte discursiva da formacao geral}

\item{media_nt_fg_d1_ct_comp:}{Media de Nota de Conteudo da questao 1 da parte discursiva da formacao geral}

\item{media_nt_fg_d2_pt_comp:}{Media de Nota de Lingua Portuguesa da questao 2 da parte discursiva da formacao geral}

\item{media_nt_fg_d2_ct_comp:}{Media de Nota de Conteudo da questao 2 da parte discursiva da formacao geral}

\item{media_nt_ce_d1_comp:}{Media de Nota da questao 1 da parte discursiva do componente especifico - Convertida para escala de 0 a 100}

\item{media_nt_ce_d2_comp:}{Media de Nota da questao 2 da parte discursiva do componente especifico - Convertida para escala de 0 a 100}

\item{media_nt_ce_d3_comp:}{Media de Nota da questao 3 da parte discursiva do componente especifico - Convertida para escala de 0 a 100}
}
\description{
Funcao para gerar os graficos do ENADE
}
\author{
Vitor Almeida - vitoralm@gmail.com
}

