% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geraGraficosRespostasCorretas.R
\name{geraGraficosRespostasCorretas}
\alias{geraGraficosRespostasCorretas}
\title{gera grafico medias}
\usage{
geraGraficosRespostasCorretas(per_comp, per_a, per_b, per_c, per_d, per_e,
  per_n)
}
\arguments{
\item{per_comp:}{Array com resposta das categorias escolhidas}

\item{per_a:}{Respostas item A}

\item{per_b:}{Respostas item B}

\item{per_c:}{Respostas item C}

\item{per_d:}{Respostas item D}

\item{per_e:}{Respostas item E}

\item{per_n:}{Respostas Nao Preenchidas}
}
\description{
Funcao para gerar os graficos do ENADE
}
\author{
Vitor Almeida - vitoralm@gmail.com
}

