% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geraIndicadores.R
\name{geraIndicadores}
\alias{geraIndicadores}
\title{gera grafico medias}
\usage{
geraIndicadores(geraGraficosMedias = FALSE,
  geraGraficosRespostasCorretas = FALSE, co_ies = c(1556, 583, 1895),
  co_uf_curso = 23, co_catad = 10005)
}
\arguments{
\item{geraGraficosMedias:}{Boleano para gerar Graficos Medias. Por Padrao False}

\item{geraGraficosRespostasCorretas:}{Boleano para gerar graficos Respostas Corretas. Por Padrao False}

\item{co_ies:}{Codigo da Instituicao de Ensino Superior. Por padrao c(1556, 583, 1895)        #UNI7, UFC, UniChristus}

\item{co_uf_curso:}{UF do Curso. Por padrao CE}

\item{co_catad:}{Categoria da Instituicao. Por padrao: Instituicao Privada sem Fins Lucrativos.}
}
\description{
Funcao para gerar os graficos do ENADE
}
\author{
Vitor Almeida - vitoralm@gmail.com
}

