% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduzDataset.R
\name{reduzDataset}
\alias{reduzDataset}
\title{reduz dataset removendo informacoes que nao serao uteis}
\usage{
reduzDataset()
}
\description{
Funcao para reduzir o tamanho do Dataset original
}
\author{
Rodrigo Almeida - rodrigo.almeida@gmail.com
}

