% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geraIndicadores.R
\name{geraIndicadores}
\alias{geraIndicadores}
\title{gera grafico medias}
\usage{
geraIndicadores(geraGraficosMedias = FALSE,
  geraGraficosRespostasCorretas = FALSE)
}
\arguments{
\item{geraGraficosMedias:}{Boleano para gerar Graficos Medias}

\item{geraGraficosRespostasCorretas:}{Boleano para gerar graficos Respostas Corretas}
}
\description{
Funcao para gerar os graficos do ENADE
}
\author{
Vitor Almeida - vitoralm@gmail.com
}

